% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio.R
\name{drawio}
\alias{drawio}
\title{Graphics device for DrawIO output}
\usage{
drawio(filename = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{filename}{If given, write ascii to this file, otherwise write to console.}

\item{width, height}{dimensions of text output (in characters). Default: NULL
(auto-detect)}

\item{...}{other parameters passed to the rdevice}
}
\description{
Graphics primitives will be rendered into DrawIO format.
}
\details{
Uses \code{devout::rdevice()}.
}
