% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio-callback.R
\name{drawio_callback}
\alias{drawio_callback}
\title{drawio callback for the rdevice}
\usage{
drawio_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device function call}

\item{args}{arguments to device function call}

\item{state}{list of rdata, dd and gc. Some or all of which may be NULL}
}
\description{
drawio callback for the rdevice
}
