% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio-callback.R
\name{drawio_metricInfo}
\alias{drawio_metricInfo}
\title{Determine the font metrics given the current drawing state}
\usage{
drawio_metricInfo(args, state)
}
\arguments{
\item{args}{arguments to this particular method}

\item{state}{full state of current drawing environment. This includes the
graphics context (\code{state$gc)}) and any R state needed for the device
(\code{state$rdata}).}
}
\description{
Determine the font metrics given the current drawing state
}
