% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio-callback.R
\name{v2hex}
\alias{v2hex}
\title{Convert a 4 element RGBA numeric vector into a colour for drawio}
\usage{
v2hex(vec)
}
\arguments{
\item{vec}{4 element numeric vector containing RGBA values in range [0,255]}
}
\description{
If alpha is zero then return none, otherwise return the hex colour without
the alpha component. Alpha is handled separately in drawio elements
}
